// CourseAdminSystem/Frontend/CourseAdminSystemAngular/src/app/model/ashit.ts
import { Shitter } from './shitter';
import { Toilet } from './toilet';

export interface AShit {
  ShitID?: number; //adding the ? to make it optional, as it will be auto-generated by the backend
  Shitterid?: number;
  Toiletid?: number;
  Time?: Date;
  Rating?: number;
  Review?: string;
  TheShitter?: Shitter;
  TheToilet?: Toilet;
}